refactor(compiler): move test helper to test file (#32937)
The `processI18nMeta()` function is only called from
a single test, so let's move it there to keep the main
source simpler.

PR Close #32937
